Split BSAReader into multiple files
