+++ Compression.BSA/BSA/Builder/BSABuilder.cs
