+++ Compression.BSA/BSA/VersionType.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace Compression.BSA
+{
+    public enum VersionType : uint
+    {
+        TES4 = 0x67,
+        FO3 = 0x68, // FO3, FNV, TES5
+        SSE = 0x69,
+        FO4 = 0x01,
+        TES3 = 0xFF // Not a real Bethesda version number
+    }
+
+}
